Refactor schema generator to entirely client-side execution without audit log
- Moves randomization logic from `server.ts` into `RandomizationService`.
- Removes server-side API endpoints (`/api/randomize`, `/api/audit-logs`) and related interfaces.
- Completely deletes the `AuditLogModalComponent` and removes UI references to it in `GeneratorComponent`.
- Updates `AboutComponent` with explicit disclaimer that the zero-trust application is not 21 CFR Part 11 compliant, requires users to maintain generated code records, and should not be used in production.
- Ensures all tests pass by providing the router in `app.spec.ts`.
