+++ src/app/app.spec.ts
@@ -1,10 +1,13 @@
 import {TestBed} from '@angular/core/testing';
+import { provideRouter } from '@angular/router';
 import {App} from './app';
+import { routes } from './app.routes';
 
 describe('App', () => {
   beforeEach(async () => {
     await TestBed.configureTestingModule({
       imports: [App],
+      providers: [provideRouter(routes)]
     }).compileComponents();
   });
